feat: websocket-polyfill.js resource
